+++ app/src/main/java/de/markusfisch/android/pielauncher/widget/AppPieView.java
@@ -31,6 +31,7 @@ import java.util.Collections;
 import java.util.List;
 
 import de.markusfisch.android.pielauncher.R;
+import de.markusfisch.android.pielauncher.activity.SettingsActivity;
 import de.markusfisch.android.pielauncher.app.PieLauncherApp;
 import de.markusfisch.android.pielauncher.content.AppMenu;
 import de.markusfisch.android.pielauncher.graphics.CanvasPieMenu;
@@ -64,19 +65,18 @@ public class AppPieView extends View {
 	private final ArrayList<AppMenu.Icon> backup = new ArrayList<>();
 	private final ArrayList<AppMenu.Icon> ungrabbedIcons = new ArrayList<>();
 	private final Paint paintActive = new Paint(Paint.FILTER_BITMAP_FLAG);
-	private final Paint paintDeactive = new Paint(Paint.FILTER_BITMAP_FLAG);
 	private final TextPaint paintText = new TextPaint(Paint.ANTI_ALIAS_FLAG);
 	private final Point touch = new Point();
 	private final Ripple ripple = new Ripple();
 	private final Rect drawRect = new Rect();
-	private final Rect iconAddRect = new Rect();
+	private final Rect iconStartRect = new Rect();
+	private final Rect iconCenterRect = new Rect();
+	private final Rect iconEndRect = new Rect();
 	private final Bitmap iconAdd;
-	private final Rect iconRemoveRect = new Rect();
 	private final Bitmap iconRemove;
-	private final Rect iconInfoRect = new Rect();
-	private final Bitmap iconInfo;
-	private final Rect iconDoneRect = new Rect();
+	private final Bitmap iconDetails;
 	private final Bitmap iconDone;
+	private final Bitmap iconPreferences;
 	private final Bitmap iconLaunchFirst;
 	private final String numberOfIconsTip;
 	private final String dragToOrderTip;
@@ -133,18 +133,18 @@ public class AppPieView extends View {
 		dragToOrderTip = context.getString(R.string.tip_drag_to_order);
 		pinchZoomTip = context.getString(R.string.tip_pinch_zoom);
 
-		paintDeactive.setAlpha(40);
 		paintText.setColor(res.getColor(R.color.text_color));
 		paintText.setTextAlign(Paint.Align.CENTER);
 		paintText.setTextSize(14f * sp);
 		textHeight = paintText.descent() - paintText.ascent();
 		textOffset = (textHeight / 2) - paintText.descent();
-		translucentBackgroundColor = res.getColor(R.color.background_ui);
+		translucentBackgroundColor = res.getColor(R.color.bg_ui);
 
 		iconAdd = getBitmapFromDrawable(res, R.drawable.ic_add);
 		iconRemove = getBitmapFromDrawable(res, R.drawable.ic_remove);
-		iconInfo = getBitmapFromDrawable(res, R.drawable.ic_info);
+		iconDetails = getBitmapFromDrawable(res, R.drawable.ic_details);
 		iconDone = getBitmapFromDrawable(res, R.drawable.ic_done);
+		iconPreferences = getBitmapFromDrawable(res, R.drawable.ic_preferences);
 		iconLaunchFirst = getBitmapFromDrawable(res,
 				R.drawable.ic_launch_first);
 		iconLaunchFirstHalf = iconLaunchFirst.getWidth() >> 1;
@@ -490,14 +490,11 @@ public class AppPieView extends View {
 					return;
 				}
 				initRipple();
-				longPressRunnable = new Runnable() {
-					@Override
-					public void run() {
-						performHapticFeedback(
-								HapticFeedbackConstants.LONG_PRESS);
-						addIconInteractively(appIcon);
-						longPressRunnable = null;
-					}
+				longPressRunnable = () -> {
+					performHapticFeedback(
+							HapticFeedbackConstants.LONG_PRESS);
+					addIconInteractively(appIcon);
+					longPressRunnable = null;
 				};
 				postDelayed(longPressRunnable, minLongPressDuration);
 			}
@@ -513,12 +510,9 @@ public class AppPieView extends View {
 			private void initRipple() {
 				cancelRipple();
 				final Point at = new Point(touch.x, touch.y + getScrollY());
-				rippleRunnable = new Runnable() {
-					@Override
-					public void run() {
-						ripple.set(at.x, at.y);
-						invalidate();
-					}
+				rippleRunnable = () -> {
+					ripple.set(at.x, at.y);
+					invalidate();
 				};
 				postDelayed(rippleRunnable, tapTimeout);
 			}
@@ -536,17 +530,14 @@ public class AppPieView extends View {
 				cancelPerformAction();
 				final boolean wasTap = isTap(event, tapTimeout);
 				final Point at = new Point(touch.x, touch.y);
-				performActionRunnable = new Runnable() {
-					@Override
-					public void run() {
-						v.performClick();
-						if (performAction(v.getContext(), at, wasTap)) {
-							v.performHapticFeedback(HAPTIC_FEEDBACK_CONFIRM);
-						}
-						performActionRunnable = null;
-						hidePieMenu();
-						invalidate();
+				performActionRunnable = () -> {
+					v.performClick();
+					if (performAction(v.getContext(), at, wasTap)) {
+						v.performHapticFeedback(HAPTIC_FEEDBACK_CONFIRM);
 					}
+					performActionRunnable = null;
+					hidePieMenu();
+					invalidate();
 				};
 				// Wait a short time before performing the action
 				// because some touch screen drivers send
@@ -582,9 +573,8 @@ public class AppPieView extends View {
 	}
 
 	private void layoutEditorControls(boolean portrait) {
-		Bitmap[] icons = new Bitmap[]{iconAdd, iconRemove, iconInfo, iconDone};
-		Rect[] rects = new Rect[]{iconAddRect, iconRemoveRect, iconInfoRect,
-				iconDoneRect};
+		Bitmap[] icons = new Bitmap[]{iconAdd, iconPreferences, iconDone};
+		Rect[] rects = new Rect[]{iconStartRect, iconCenterRect, iconEndRect};
 		int length = icons.length;
 		int totalWidth = 0;
 		int totalHeight = 0;
@@ -684,7 +674,10 @@ public class AppPieView extends View {
 		} else if (mode == MODE_LIST && wasTap) {
 			return performListAction(context, at);
 		} else if (mode == MODE_EDIT) {
-			return performEditAction(context);
+			boolean result = performEditAction(context);
+			grabbedIcon = null;
+			PieLauncherApp.appMenu.updateSmoothing();
+			return result;
 		}
 		return false;
 	}
@@ -714,43 +707,34 @@ public class AppPieView extends View {
 	}
 
 	private boolean performEditAction(Context context) {
-		boolean successful = false;
-		if (iconAddRect.contains(touch.x, touch.y)) {
-			if (grabbedIcon != null) {
-				rollback();
-			} else {
+		if (iconStartRect.contains(touch.x, touch.y)) {
+			if (grabbedIcon == null) {
 				((Activity) context).onBackPressed();
-			}
-			successful = true;
-		} else if (iconRemoveRect.contains(touch.x, touch.y)) {
-			ripple.set(touch);
-			if (grabbedIcon != null) {
+			} else {
+				ripple.set(touch);
 				PieLauncherApp.appMenu.icons.remove(grabbedIcon);
 				// Undo any rotation if the menu has not otherwise changed.
 				if (sameOrder(backup, PieLauncherApp.appMenu.icons)) {
 					rollback();
 				}
 			}
-			successful = true;
-		} else if (iconInfoRect.contains(touch.x, touch.y)) {
-			ripple.set(touch);
-			if (grabbedIcon != null) {
+			return true;
+		} else if (iconCenterRect.contains(touch.x, touch.y) &&
+				grabbedIcon == null) {
+			SettingsActivity.start(context);
+			return true;
+		} else if (iconEndRect.contains(touch.x, touch.y)) {
+			if (grabbedIcon == null) {
+				endEditMode();
+			} else {
+				ripple.set(touch);
 				rollback();
 				PieLauncherApp.appMenu.launchAppInfo(context,
 						(AppMenu.AppIcon) grabbedIcon);
 			}
-			successful = true;
-		} else if (iconDoneRect.contains(touch.x, touch.y)) {
-			if (grabbedIcon != null) {
-				rollback();
-			} else {
-				endEditMode();
-			}
-			successful = true;
+			return true;
 		}
-		grabbedIcon = null;
-		PieLauncherApp.appMenu.updateSmoothing();
-		return successful;
+		return false;
 	}
 
 	private void rollback() {
@@ -864,11 +848,19 @@ public class AppPieView extends View {
 	private void drawEditor(Canvas canvas) {
 		canvas.drawColor(translucentBackgroundColor, PorterDuff.Mode.SRC);
 		boolean hasIcon = grabbedIcon != null;
-		drawTip(canvas, getTip(hasIcon));
-		drawIcon(canvas, iconAdd, iconAddRect, !hasIcon);
-		drawIcon(canvas, iconRemove, iconRemoveRect, hasIcon);
-		drawIcon(canvas, iconInfo, iconInfoRect, hasIcon);
-		drawIcon(canvas, iconDone, iconDoneRect, !hasIcon);
+		// Only draw tips in portrait orientation.
+		// There's probably not enough space in landscape.
+		if (canvas.getWidth() < canvas.getHeight()) {
+			drawTip(canvas, getTip(hasIcon));
+		}
+		if (hasIcon) {
+			drawIcon(canvas, iconRemove, iconStartRect);
+			drawIcon(canvas, iconDetails, iconEndRect);
+		} else {
+			drawIcon(canvas, iconAdd, iconStartRect);
+			drawIcon(canvas, iconPreferences, iconCenterRect);
+			drawIcon(canvas, iconDone, iconEndRect);
+		}
 		int centerX = viewWidth >> 1;
 		int centerY = viewHeight >> 1;
 		setCenter(centerX, centerY);
@@ -973,10 +965,8 @@ public class AppPieView extends View {
 		}
 	}
 
-	private void drawIcon(Canvas canvas, Bitmap icon, Rect rect,
-			boolean active) {
-		canvas.drawBitmap(icon, null, rect,
-				active ? paintActive : paintDeactive);
+	private void drawIcon(Canvas canvas, Bitmap icon, Rect rect) {
+		canvas.drawBitmap(icon, null, rect, paintActive);
 	}
 
 	private static float distSq(Point a, Point b) {